chore(weave_ts): Move call-based tracing processor for OpenAI Agents SDK into its own module.
